GuessworkQuantumSideInfo

GuessworkQuantumSideInfo.BB84_states

GuessworkQuantumSideInfo.guesswork

GuessworkQuantumSideInfo.guesswork_lower_bound

GuessworkQuantumSideInfo.pmfN

GuessworkQuantumSideInfo.randdm

GuessworkQuantumSideInfo.randprobvec

GuessworkQuantumSideInfo.randpure

Main functions

GuessworkQuantumSideInfo.guesswork — Function

guesswork(
    p::AbstractVector{numeric_type},
    ρBs::AbstractVector{<:AbstractMatrix};
    solver,
    K::Integer = length(p),
    c = numeric_type[1:K..., 5_000],
    dual::Bool = false,
    remove_repetition::Bool = true,
    povm_outcomes = make_povm_outcomes(length(p), K, remove_repetition),
    verbose::Bool = true,
    debug::Bool = false,
)

Compute the guesswork.

GuessworkQuantumSideInfo.guesswork_lower_bound — Function

guesswork_lower_bound(
    p::AbstractVector{T},
    ρBs::AbstractVector{<:AbstractMatrix};
    solver,
    c = T[1:length(p)..., 10_000],
    verbose::Bool = false,
)

Computes a lower bound to the optimal expected number of guesses by relaxing the problem. For J states, only needs J^2 PSD variables subject to two linear constraints.

Gives e.g. 1.5 instead of 1.71 for BB84(1), and 2.5 for BB84(2), providing the current tightest lower bound to BB84(2) (we can get an upper bound of ~3.7 via dual relaxations). Doesn't seem like a great lower bound, but it does seem to perform better on random inputs (e.g. giving 1.91 instead of 1.98 in a four random qubit example.)

Missing docstring.

Missing docstring for GuessworkQuantumSideInfo.guesswork_upper_bound. Check Documenter's build log for details.

GuessworkQuantumSideInfo.pmfN — Function

pmfN(data; tol = 1e-5) -> Vector

Compute the probability mass function for the number of guesses N, given a strategy. The nth entry of the output vector vec gives the probability for guessing the correct answer on the nth try, for n = 1 : K. The last entry gives the probability of never guessing the correct answer.

data should be a NamedTuple with entries for p, ρBs, Es, K, and povm_outcomes

tol provides a tolerance above which to warn about imaginary or negative probabilities.

Quantum states

GuessworkQuantumSideInfo.randdm — Function

randdm(T, d)

Generates a density matrix of dimension d at random.

GuessworkQuantumSideInfo.randpure — Function

randpure(T, d)

Generates a pure state density matrix of dimension d at random.

GuessworkQuantumSideInfo.randprobvec — Function

randprobvec([T=Float64], d)

Generates points of type T, uniformly at random on the standard d-1 dimensional simplex using an algorithm by Smith and Tromble.

GuessworkQuantumSideInfo.BB84_states — Function

BB84_states([T::Type = Float64])

Generates the BB84 states for use in guesswork or other functions. The numeric type can be optionally specified by the argument.