GuessworkQuantumSideInfo

This is a package accompanying the preprint Guesswork with Quantum Side Information.

See the Examples for some examples (or just below for a quick example), Computing the guesswork to high precision for an example solving a problem with a high-precision SDP solver, Using a mixed-integer SDP to find extremal strategies for an example using a mixed-integer SDP, or below for the documentation of the functions provided by this package.

Quick example

Consider one party Alice who draws a random number in the set [1,2,3,4] uniformly at random. If she draws 1 she sends another party, Bob, the quantum state |0⟩; if she draws 2, she sends |1⟩, if she draws 3 she sends |-⟩, and finally if she draws 4, she sends |+⟩. Bob, knowing this general procedure but not which number Alice drew, aims to guess the value Alice drew by performing experiments on the quantum state he was given. The average number of guesses Bob needs in order to get the right answer, minimized over all quantum strategies, is the so-called guesswork with quantum side information. This package provides a means to compute this.

julia> using GuessworkQuantumSideInfo, SCS

julia> p = [0.25, 0.25, 0.25, 0.25];

julia> ketzero = ket(1, 2);

julia> ketone = ket(2, 2);

julia> ketminus = (ket(1, 2) - ket(2,2))/sqrt(2);

julia> ketplus = (ket(1, 2) + ket(2,2))/sqrt(2);

julia> ρBs = dm.([ ketzero, ketone, ketminus, ketplus  ])
4-element Array{Array{Complex{Float64},2},1}:
 [1.0 + 0.0im 0.0 + 0.0im; 0.0 + 0.0im 0.0 + 0.0im]
 [0.0 + 0.0im 0.0 + 0.0im; 0.0 + 0.0im 1.0 + 0.0im]
 [0.4999999999999999 + 0.0im -0.4999999999999999 - 0.0im; -0.4999999999999999 + 0.0im 0.4999999999999999 + 0.0im]
 [0.4999999999999999 + 0.0im 0.4999999999999999 + 0.0im; 0.4999999999999999 + 0.0im 0.4999999999999999 + 0.0im]

julia> output = guesswork(p, ρBs; solver = SCSSolver(verbose=false));

julia> output.optval
1.7094231882324955

Guesswork functions

GuessworkQuantumSideInfo.guesswork — Function

guesswork(
    p::AbstractVector{T},
    ρBs::AbstractVector{<:AbstractMatrix};
    solver,
    K::Integer = length(p),
    c = T[1:K..., 5_000],
    dual::Bool = false,
    remove_repetition::Bool = true,
    povm_outcomes = make_povm_outcomes(length(p), K, remove_repetition),
    verbose::Bool = true,
)

Computes the guesswork for the c-q state specified by a probability vector p, giving the distribution X, and ρBs, giving the associated quantum states.

The keyword arguments are as follows:

solver is the only required keyword argument; an SDP solver such as SCS or MOSEK must be passed.

K corresponds to the maximum number of allowed guesses. The number of variables in the primal SDP (and the number of constraints in the dual SDP) scales as length(p)^K.

c may be given a custom cost vector. If K < length(p), then c should be of length K+1. The last entry, c[K+1], corresponds to the cost of not guessing the correct answer within K guesses.

dual is a boolean variable indicating whether the primal or dual optimization problem should be solved.

remove_repetition is a boolean variable defaulting to true, indicating whether repeated guesses of the same value should be removed; as long as c is increasing, this decreases the size of the SDP without affecting the optimal value.

povm_outcomes should be an iterator (or vector) corresponding to the possible guessing orders. This defaults to all subsets of length K of 1:length(p) without repetition.

verbose is a boolean which indicates if warnings should be printed when the problem is not solved optimally.

GuessworkQuantumSideInfo.guesswork_lower_bound — Function

guesswork_lower_bound(
    p::AbstractVector{T},
    ρBs::AbstractVector{<:AbstractMatrix};
    solver,
    c = T[1:length(p)..., 10_000],
    verbose::Bool = false,
)

See guesswork for the meaning of the arguments. Computes a lower bound to the optimal expected number of guesses by solving a relaxed version of the primal SDP. For J states, only needs J^2 PSD variables subject to two linear constraints.

GuessworkQuantumSideInfo.guesswork_upper_bound — Function

guesswork_upper_bound(
    p::AbstractVector{T},
    ρBs::AbstractVector{<:AbstractMatrix};
    make_solver,
    c::AbstractVector = T.(1:length(p)),
    max_retries = 50,
    max_time = Inf,
    num_constraints = Inf,
    verbose::Bool = false,
    num_steps_per_SA_run::Integer = length(p)^2 * 500,
) where {T<:Number} -> NamedTuple

Computes an upper bound to the guesswork problem associated to the c-q state specified by p and ρBs, as in guesswork. A custom cost vector c may be optionally passed. If the keyword argument verbose is set to true, information is printed about each iteration of the algorithm.

The keyword argument make_solver is required, and must pass a function that creates a solver instances. For example, instead of passing SCSSolver(), pass () -> SCSSolver(). This is needed because the algorithm used in guesswork_upper_bound solves a sequence of SDPs, not just one.

The algorithm has three termination criteria which are controlled by keyword arguments. The algorithm stops when any of the following occur:

max_retries simulated annealing attempts fail to find a violated constraint.

num_constraints constraints have been added to the dual SDP

The total runtime of the algorithm is projected to exceed max_time on the next iteration.

By default, max_retries is set to 50, while num_constraints and max_time are set to infinity.

Lastly, the keyword argument num_steps_per_SA_run controls the runtime of the simulated annealing algorithm. Increase num_steps_per_SA_run to search longer for a violated constraint within a given simulated annealing run.

GuessworkQuantumSideInfo.guesswork_ellipsoid — Function

guesswork_ellipsoid(p, ρBs; nl_solver, kwargs...) -> NamedTuple

Solves the guesswork SDP via a custom implementation of the ellipsoid algorithm. When provided a global optimizer for the nl_solver keyword (e.g. via EAGO.jl), this computes the value guesswork SDP. Keyword arguments are:

nl_solver: choice of nonlinear optimizer to use

c::AbstractVector = T.(1:length(p)): choice of cost vector

deepcut = true: whether to use "deep cuts" (can be faster)

tol = 1e-3: solve tolerance (final value should be within tol of true solution)

max_time::TimePeriod: provide an approximate upper limit on the duration of the solve (defaults to unlimited)

verbose::Bool = true: whether to print timing information every so often to the terminal

timer_log_interval::Millisecond = Millisecond(1)*1e4: how often to print timing information

logger = nothing: provide a logger (such as via TensorBoardLogger.jl) to log information during a run.

normal_cone_tol = 0.0: absolute tolerance to decide if a number is negative for the purpose of computing an element in the normal cone

perm_tol = 1e-4: a relative tolerance to decide if a number is 1 for the purpose of finding permutations in the support of a doubly stochastic matrix

init_noise = 1e-6: an amount of noise to add to the initial point to reduce symmetries

trace = true: whether or not to store a trace of the center and shape of the ellipsoid, along with other parameters, at each step

max_SA_retries = 2: a parameter to decide how many times to run a simulated annealing algorithm to find cuts before resorting to global optimization

num_steps_per_SA_run::Integer = length(p)^2 * 500: how many steps to perform during each simulated annealing run

mutate! = rand_rev!: mutation function for updating during the simulated annealing run

x = nothing: initial center point of the ellipse (automatically chosen by default_init if left unspecified)

P = nothing: initial shape of the ellipse (automatically chosen by default_init if left unspecified)

To aid in resuming a previous run (after hitting a time limit or tightening tol), the following parameters may be passed:

x_best = nothing:

timer = TimerOutput():

cuts = Vector{Int}[]:

iter = Ref(1):

f_best = Ref(T(Inf)):

tracelog = []:

If results = guesswork_ellipsoid(p, ρBs; nl_solver = ...), then results.prob holds the previous solution object. This run may be continued by ellipsoid_algorithm!(results.prob; tol=...) or fields may be passed here, e.g. guesswork_ellipsoid(p, ρBs; x_best = results.prob.x_best, ...), or to the EllipsoidProblem constructor, e.g. EllipsoidProblem(x_best = results.prob.x_best, ...).

Quantum states

GuessworkQuantumSideInfo.ket — Function

ket([T = Float64], i::Integer, d::Integer) -> SparseVector{Complex{T}}

Create a vector representing the ith computational basis vector in dimension d.

Example

julia> ket(1,2)
2-element SparseVector{Complex{Float64},Int64} with 1 stored entry:
  [1]  =  1.0+0.0im

julia> collect(ans)
2-element Array{Complex{Float64},1}:
 1.0 + 0.0im
 0.0 + 0.0im

GuessworkQuantumSideInfo.bra — Function

bra([T = Float64], i::Integer, d::Integer) -> SparseVector{Complex{T}}'

Create a dual vector representing the bra associated to ith computational basis vector in dimension d.

Example

julia> bra(1,2)
1×2 LinearAlgebra.Adjoint{Complex{Float64},SparseVector{Complex{Float64},Int64}}:
 1.0-0.0im  0.0-0.0im

julia> collect(ans)
1×2 Array{Complex{Float64},2}:
 1.0-0.0im  0.0-0.0im

GuessworkQuantumSideInfo.BB84_states — Function

BB84_states([T::Type = Float64])

Generates the BB84 states |0⟩, |1⟩, |-⟩, and |+⟩, for use in guesswork or other functions. The numeric type can be optionally specified by the argument.

GuessworkQuantumSideInfo.iid_copies — Function

iid_copies(ρBs::AbstractVector{<:AbstractMatrix}, n::Integer) -> Vector{Matrix}

Create a vector of all states of the form $ρ_1 \otimes \dotsm \otimes ρ_n$ where the $ρ_i$ range over the set ρBs.

GuessworkQuantumSideInfo.randdm — Function

randdm([T = Float64], d)

Generates a density matrix with numeric type Complex{T}, of dimension d at random.

Example

julia> randdm(2)
2×2 Array{Complex{Float64},2}:
 0.477118+0.0im        0.119848-0.0371569im
 0.119848+0.0371569im  0.522882+0.0im      

GuessworkQuantumSideInfo.randprobvec — Function

randprobvec([T=Float64], d)

Generates points of type T, uniformly at random on the standard d-1 dimensional simplex using an algorithm by Smith and Tromble.

Example

julia> randprobvec(3)
3-element Array{Float64,1}:
 0.24815974900033688
 0.17199716455672287
 0.5798430864429402 

Utilities

GuessworkQuantumSideInfo.pmfN — Function

pmfN(data; tol = 1e-5) -> Vector

Compute the probability mass function for the number of guesses N, given a strategy. The nth entry of the output vector gives the probability for guessing the correct answer on the nth try, for n = 1 : K. If the number of allowed guesses, K, is smaller than length(p), then there is an additional last entry which gives the probability of never guessing the correct answer.

data should be a NamedTuple with entries for p, ρBs, Es, K, and povm_outcomes

tol provides a tolerance above which to warn about imaginary or negative probabilities.

Ellipsoid algorithm functions

GuessworkQuantumSideInfo.default_init — Function

default_init(p, ρBs, c, dB, init_noise) -> NamedTuple

Choose an initial center x and shape P of the ellipse such that the solution is guaranteed to be inside the ellipse. The initial point is chosen to correspond to c[1]*ρB + noise where ρB is the average state, and noise is a small perturbation whose norm is governed by init_noise.

GuessworkQuantumSideInfo.EllipsoidProblem — Type

struct EllipsoidProblem

Stores settings and all mutable state of the ellipsoid method. This allows, e.g.

results = guesswork_ellipsoid(p, ρBs; tol=1e-3, nl_solver = ...)
# inspect `out.optval`, etc
# Continue to solve with a tighter solution tolerance:
results2 = ellipsoid_algorithm!(results.prob; tol=1e-4)

See guesswork_ellipsoid for the possible keyword arguments for constructing an EllipsoidProblem, which may be constructed by e.g. EllipsoidProblem(p, ρBs; nl_solver = ..., kwargs...).

GuessworkQuantumSideInfo.invherm — Function

invherm(M::AbstractMatrix) -> Vector

Creates a real vector v representation of a complex Hermitian matrix M. These are related by a linear isometry (with respect to the 2-norm). See also herm.

Example

M = Hermitian(rand(4,4) + im*rand(4,4))
v = invherm(M)

norm(v, 2) ≈ norm(M, 2)

GuessworkQuantumSideInfo.herm — Function

herm(M::AbstractVector) -> Hermitian

Creates the complex Hermitian matrix M represented by the real vector v. See also invherm.

Example

v = rand(16)
M = herm(M)

norm(v, 2) ≈ norm(M, 2)

GuessworkQuantumSideInfo.ellipsoid_algorithm! — Function

ellipsoid_algorithm!(f::EllipsoidProblem; tol=nothing, max_time=nothing)

Run the ellipsoid algorithm on the EllipsoidProblem f. Optionally, pass the keyword argument tol or max_time to update these parameters of f before running.

GuessworkQuantumSideInfo.PermutationIterator — Type

PermutationIterator(D::AbstractMatrix; timer=nothing, rtol=1e-6)

Iterates over permutations and weights to form a decomposition of a doubly stochastic matrix D, known as the Birkhoff–von Neumann decomposition.

Optionally, specify the keyword arguments

timer: a TimerOutputs.jl timer

rtol: (relative tolerance) for comparing α_i to 1 to see if all the permutations have been found.

Example

# Construct a `d` by `d` doubly stochastic matrix
d = 4
α_init = rand(5)
α_init = α_init / sum(α_init)
D = sum( α_init[i]*I(d)[randperm(d), :] for i = 1:5 )

# Reconstruct it from permutations
D_reconstruct = zero(D)
for (π_i, α_i) in PermutationIterator(D)
    P_i = I(d)[π_i, :]
    global D_reconstruct += α_i * P_i
end
@test D_reconstruct ≈ D